related work: add 5gbasechecker for ota fuzzing
This is one of the most recent works and orthogonal to the re-hosting.
Nonetheless, it contains very interesting insights that can be combined
with Re-hosting: e.g. "differential fuzzing" and more exhaustive FSM
creation. In the end it is not necessary to restrict to 40s/req as for a
manual black-box method.