equivalence-checking: add the algorithm for the w-method
The algorithm provides a (non-optimal) way to do equivalence-checking
for a given specification (in active learning: the current hypothesis).

The idea is to find all the distinguishing words (set: $W$) and starting
states (reached by words in the set $P$), that covers the transitions of
the testing tree.

Together, all the test cases cover every word, that might show any form
of faulty behavior.

evaluate those from. If all

Signed-off-by: Eduard Vlad <eduard.vlad@rwth-aachen.de>